Add reboot functionality and update variable types for irrigation timing

+++ src/globals.h
@@ -16,7 +16,7 @@ extern RTC_DS3231 rtc;
 
 extern IrrigationSchedules schedules;
 extern bool irrigationScheduleEnabled;
-extern int irrigationManualEnd;
+extern unsigned long irrigationManualEnd;
 extern bool isValidForDayRun(DateTime date, int dayToRun);
 extern int displayNetworkActivity;
 extern int displayOutChange;
